migrate test for component Filters to testing-library
